+++ src/Events/ScheduledReportDispatched.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace MBLSolutions\Report\Events;
+
+use MBLSolutions\Report\Models\ScheduledReport;
+
+class ScheduledReportDispatched
+{
+
+    public ScheduledReport $schedule;
+
+    public function __construct(ScheduledReport $schedule)
+    {
+        $this->schedule = $schedule;
+    }
+
+}